neuralfetch: split CLI into commands/ package and utils/ library
Restructures the extracted CLI into per-file argparse commands and
consolidates the library, replacing the single cli/ package.

commands/ (CLI glue; argparse subparser aggregation keeps --help automatic):
- commands/__init__.py: dispatcher + command registry; commands/__main__.py
- commands/download.py, study_info.py, export_bids.py: one file per subcommand
  (download / study-info / export-bids), each with NAME/HELP/add_arguments/run;
  heavy imports stay inside run() so subparser registration is cheap

utils/ (library; neuralfetch.utils.* import paths preserved via __init__ re-exports):
- split utils.py into utils/study_info.py (root_study_folder, compute_study_info,
  format_study_info, update_source_info) and utils/data.py (add_sentences,
  scan_files, ensure_imagenet22k, download_things_images)
- cli/bids_exporter.py -> utils/bids.py (restores the neuralfetch.utils.bids path)
- cli/download.py -> utils/runner.py; tests moved to utils/test_bids.py, test_runner.py

- [project.scripts] neuralfetch = "neuralfetch.commands:main"
- docs reference points at neuralfetch.utils.runner / neuralfetch.utils.bids
