--- neuralfetch-repo/neuralfetch/cli/__init__.py
@@ -1,11 +1,0 @@
-# Copyright (c) Meta Platforms, Inc. and affiliates.
-# All rights reserved.
-#
-# This source code is licensed under the license found in the
-# LICENSE file in the root directory of this source tree.
-
-"""neuralfetch command-line interface package."""
-
-from neuralfetch.cli.main import main
-
-__all__ = ["main"]
