+++ neuralfetch-repo/neuralfetch/commands/__main__.py
@@ -4,9 +4,9 @@
 # This source code is licensed under the license found in the
 # LICENSE file in the root directory of this source tree.
 
-"""Entry point for ``python -m neuralfetch.cli``; delegates to :mod:`neuralfetch.cli.main`."""
+"""Entry point for ``python -m neuralfetch.commands``; delegates to :func:`neuralfetch.commands.main`."""
 
-from neuralfetch.cli.main import main
+from neuralfetch.commands import main
 
 if __name__ == "__main__":
     main()
